Update and rename Lucky_number.java to Question_12.java
